Added new types.
Added types to properly reflect the Events API responses. It decodes
the arbitrary entity metadata via the EventKind type, and then
creates the unnamed struct and maps it to the entity metadata.

The tests just loop through some subtests because I was really
lazy and didn't want to write individual tests. Sorry.

Signed-off-by: Mike Lloyd <mlloyd@pivotal.io>